Move interceptor to debug
