+++ src/agent/debug/interceptor.js
